распределяет утилитарные функции по отдельным файлам
